--- app/containers/Header/data.js
@@ -1,14 +1,0 @@
-import { connect } from 'react-redux'
-import { dispatchers } from '~/stores/auth'
-
-const mapStateToProps = state => ({
-  authenticated: state.auth.authenticated
-})
-
-const mapDispatchToProps = dispatch => ({
-  actions: {
-    logout: () => dispatch(dispatchers.signOut())
-  }
-})
-
-export default comp => connect(mapStateToProps, mapDispatchToProps)(comp)
