+++ app/containers/PostInfo/index.js
@@ -1,58 +1,18 @@
-import moment from 'moment'
-import PropTypes from 'prop-types'
-import * as S from './styles'
-import connect from './data'
+import { graphql } from 'react-apollo'
+import getPostGql from './getPost.gql'
+import Feature from './feature'
 
-const PostInfo = ({ loading, Post, error }) => {
-  if (loading) {
-    return (
-      <S.Section>
-        <h1>Loading...</h1>
-      </S.Section>
-    )
-  }
+const withData = graphql(getPostGql, {
+  options: ({ postId }) => ({
+    variables: {
+      postId
+    }
+  }),
+  props: ({ data: { loading, Post, error } }) => ({
+    loading,
+    Post,
+    error
+  })
+})
 
-  if (error) {
-    console.log(error) // eslint-disable-line no-console
-    window.alert('Load error, check console') // eslint-disable-line no-alert
-    return
-  }
-
-  return (
-    <S.Section>
-      <h1>
-        {Post.title}
-      </h1>
-      <div>
-        <span>
-          ID: <b>{Post.id}</b>
-        </span>
-        <span>&nbsp;|&nbsp;</span>
-        <span>
-          Created At: {' '}
-          <b>{moment(Post.createdAt).format('DD.MM.YYYY kk:mm')}</b>
-        </span>
-      </div>
-      <p>
-        <S.A target="_blank" href={Post.url} rel="noopener noreferrer nofollow">
-          {Post.url}
-        </S.A>
-      </p>
-    </S.Section>
-  )
-}
-
-PostInfo.propTypes = {
-  loading: PropTypes.bool.isRequired,
-  Post: PropTypes.object,
-  error: PropTypes.object,
-  postId: PropTypes.string.isRequired,
-  postTitle: PropTypes.string.isRequired
-}
-
-PostInfo.defaultProps = {
-  Post: null,
-  error: null
-}
-
-export default connect(PostInfo)
+export default withData(Feature)
