+++ app/containers/PostList/index.js
@@ -1,55 +1,34 @@
-import PropTypes from 'prop-types'
-import { Link } from '~/routes'
-import PostUpvoter from '~/containers/PostUpvoter'
-import * as S from './styles'
-import connect from './data'
+import { graphql } from 'react-apollo'
+import allPostsGql from './allPosts.gql'
+import Feature from './feature'
 
-const PostList = ({
-  data: { allPosts, loading, _allPostsMeta },
-  loadMorePosts
-}) => {
-  if (allPosts && allPosts.length) {
-    const areMorePosts = allPosts.length < _allPostsMeta.count
-    return (
-      <S.Main>
-        <S.ItemList>
-          {allPosts.map((post, index) =>
-            <S.Item key={post.id}>
-              <div>
-                <S.Index>
-                  {index + 1}.{' '}
-                </S.Index>
-                <Link
-                  route="details"
-                  params={{
-                    postId: post.id,
-                    postTitle: encodeURIComponent(post.title)
-                  }}
-                  passHref
-                >
-                  <S.Title>
-                    {post.title}
-                  </S.Title>
-                </Link>
-                <PostUpvoter id={post.id} votes={post.votes} />
-              </div>
-            </S.Item>
-          )}
-        </S.ItemList>
-        {areMorePosts
-          ? <S.ShowMore onClick={() => loadMorePosts()}>
-              {loading ? 'Loading...' : 'Show More'}
-            </S.ShowMore>
-          : ''}
-      </S.Main>
-    )
-  }
-  return <S.Loading>Loading</S.Loading>
-}
+const POSTS_PER_PAGE = 10
 
-PostList.propTypes = {
-  data: PropTypes.object.isRequired,
-  loadMorePosts: PropTypes.func.isRequired
-}
+const withData = graphql(allPostsGql, {
+  options: () => ({
+    variables: {
+      skip: 0,
+      first: POSTS_PER_PAGE
+    }
+  }),
+  props: ({ data }) => ({
+    data,
+    loadMorePosts: () =>
+      data.fetchMore({
+        variables: {
+          skip: data.allPosts.length
+        },
+        updateQuery: (previousResult, { fetchMoreResult }) => {
+          if (!fetchMoreResult) {
+            return previousResult
+          }
+          return Object.assign({}, previousResult, {
+            // Append the new posts results to the old one
+            allPosts: [...previousResult.allPosts, ...fetchMoreResult.allPosts]
+          })
+        }
+      })
+  })
+})
 
-export default connect(PostList)
+export default withData(Feature)
